Moved micro QR files into qrcode directory.
Moved micro QR files into qrcode directory so that code re-use is simplified and there is no need to consider any hierarchy between micro and qr code.
