+++ core/CMakeLists.txt
@@ -210,19 +210,19 @@ set (MICROQRCODE_FILES
 )
 if (BUILD_READERS)
     set (MICROQRCODE_FILES ${MICROQRCODE_FILES}
-        src/microqrcode/MQRCornerFinder.h
-        src/microqrcode/MQRCornerFinder.cpp
-        src/microqrcode/MQRDetector.h
-        src/microqrcode/MQRDetector.cpp
-        src/microqrcode/MQRDimension.h
-        src/microqrcode/MQRFakeCenterCalculator.h
-        src/microqrcode/MQRFakeCenterCalculator.cpp
-        src/microqrcode/MQRFinderPattern.h
-        src/microqrcode/MQRFinderPattern.cpp
-        src/microqrcode/MQRFinderPatternFinder.h
-        src/microqrcode/MQRFinderPatternFinder.cpp
-        src/microqrcode/MQRFinderPatternInfo.h
-        src/microqrcode/MQRFinderPatternInfo.cpp
+        src/qrcode/MQRCornerFinder.h
+        src/qrcode/MQRCornerFinder.cpp
+        src/qrcode/MQRDetector.h
+        src/qrcode/MQRDetector.cpp
+        src/qrcode/MQRDimension.h
+        src/qrcode/MQRFakeCenterCalculator.h
+        src/qrcode/MQRFakeCenterCalculator.cpp
+        src/qrcode/MQRFinderPattern.h
+        src/qrcode/MQRFinderPattern.cpp
+        src/qrcode/MQRFinderPatternFinder.h
+        src/qrcode/MQRFinderPatternFinder.cpp
+        src/qrcode/MQRFinderPatternInfo.h
+        src/qrcode/MQRFinderPatternInfo.cpp
     )
 endif()
 
@@ -419,7 +419,7 @@ source_group (Sources FILES ${COMMON_FILES})
 source_group (Sources\\aztec FILES ${AZTEC_FILES})
 source_group (Sources\\datamatrix FILES ${DATAMATRIX_FILES})
 source_group (Sources\\maxicode FILES ${MAXICODE_FILES})
-source_group (Sources\\microqrcode FILES ${MICROQRCODE_FILES})
+source_group (Sources\\qrcode FILES ${MICROQRCODE_FILES})
 source_group (Sources\\oned FILES ${ONED_FILES})
 source_group (Sources\\oned\\rss FILES ${ONED_RSS_FILES})
 source_group (Sources\\pdf417 FILES ${PDF417_FILES})
