+++ core/src/qrcode/MQRCornerFinder.cpp
