+++ core/src/qrcode/MQRCornerFinder.h
