+++ core/src/qrcode/MQRDetector.cpp
