+++ core/src/qrcode/MQRDetector.h
