+++ core/src/qrcode/MQRDimension.h
