+++ core/src/qrcode/MQRFakeCenterCalculator.cpp
