+++ core/src/qrcode/MQRFakeCenterCalculator.h
