+++ core/src/qrcode/MQRFinderPattern.cpp
