+++ core/src/qrcode/MQRFinderPattern.h
