+++ core/src/qrcode/MQRFinderPatternFinder.cpp
