+++ core/src/qrcode/MQRFinderPatternFinder.h
