+++ core/src/qrcode/MQRFinderPatternInfo.cpp
