+++ core/src/qrcode/MQRFinderPatternInfo.h
